Add MS and Performance timers

+++ Project/Logging.h
@@ -8,7 +8,7 @@
 #define FPS_LOG_SIZE 100
 
 void log(const char file[], int line, const char* format, ...);
-void log_delta_ms(unsigned delta_ms);
+void log_delta_ms(float delta_ms);
 
 extern std::string log_string;
 extern int fps_log_index;
